# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  models.Ad:
    properties:
      adType:
        $ref: '#/definitions/models.AdType'
      createdAt:
        type: string
      id:
        type: integer
      metadata:
        $ref: '#/definitions/models.AdMetaData'
      name:
        type: string
      program:
        $ref: '#/definitions/models.Program'
      programID:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.AdMetaData:
    type: object
  models.AdType:
    enum:
    - VIDEO
    - BANNER
    - NATIVE
    type: string
    x-enum-varnames:
    - AdTypeVideo
    - AdTypeBanner
    - AdTypeNative
...
  models.Program:
    properties:
      bidPrice:
        type: number
      budget:
        type: number
      createdAt:
        type: string
      description:
        type: string
      endDate:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isVerified:
        type: boolean
      metadata:
        $ref: '#/definitions/models.ProgramMetaData'
      name:
        type: string
      startDate:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userID:
        type: integer
    type: object
  models.ProgramMetaData:
    properties:
      category:
        type: string
      keyword:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      isVerified:
        type: boolean
      metadata:
        $ref: '#/definitions/models.UserMetaData'
      name:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      userType:
        $ref: '#/definitions/models.UserType'
    type: object
  models.UserMetaData:
    type: object
  models.UserType:
    enum:
    - ADMIN
    - PUBLISHER
    - ADVERTISER
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RolePublisher
    - RoleAdvertiser
...
host: localhost:8000
info:
  contact: {}
  description: REST API for AdAli.
  title: AdAli API
  version: "1.0"
paths:
...
  /ads:
    get:
      description: |-
        Get all ads or filter and order them using query parameters.

        Filtering operators:
        :   (equals)
        !:  (not equals)
        >   (greater than)
        >=  (greater than or equal)
        <   (less than)
        <=  (less than or equal)
        ~   (contains, case-insensitive)

        Examples:
        /ads?filter=is_active:true
        /ads?filter=program_id:1
        /ads?filter=bid_price>=1000
        /ads?order_by=bid_price&sort=desc
        /ads?order_by=created_at&order_by=bid_price&sort=desc&sort=asc
      parameters:
      - collectionFormat: csv
        description: Filter expression. Can be repeated.
        in: query
        items:
          type: string
        name: filter
        type: array
      - collectionFormat: csv
        description: Fields to order by. Can be repeated.
        in: query
        items:
          type: string
        name: order_by
        type: array
      - collectionFormat: csv
        description: Sort direction for each order_by field (asc or desc). Can be
          repeated.
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Ad'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get ads
      tags:
      - Ads
    post:
      consumes:
      - application/json
      description: Create a new ad.
      parameters:
      - description: Ad
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/models.Ad'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Ad'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create ad
      tags:
      - Ads
  /ads/{id}:
    delete:
      description: Delete an ad by ID.
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete ad
      tags:
      - Ads
    put:
      consumes:
      - application/json
      description: Update an existing ad.
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated ad
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/models.Ad'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Ad'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update ad
      tags:
      - Ads
...
  /programs:
    get:
      description: |-
        Get all programs or filter and order them using query parameters.

        Filtering operators:
        :   (equals)
        !:  (not equals)
        >   (greater than)
        >=  (greater than or equal)
        <   (less than)
        <=  (less than or equal)
        ~   (contains, case-insensitive)

        Examples:
        /programs?filter=is_active:true
        /programs?filter=user_id:1
        /programs?filter=budget>=1000
        /programs?order_by=budget&sort=desc
        /programs?order_by=created_at&order_by=budget&sort=desc&sort=asc
      parameters:
      - collectionFormat: csv
        description: Filter expression. Can be repeated.
        in: query
        items:
          type: string
        name: filter
        type: array
      - collectionFormat: csv
        description: Fields to order by. Can be repeated.
        in: query
        items:
          type: string
        name: order_by
        type: array
      - collectionFormat: csv
        description: Sort direction for each order_by field (asc or desc). Can be
          repeated.
        in: query
        items:
          type: string
        name: sort
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Program'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get programs
      tags:
      - Programs
    post:
      consumes:
      - application/json
      description: Create a new program.
      parameters:
      - description: Program
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/models.Program'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Program'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create program
      tags:
      - Programs
  /programs/{id}:
    delete:
      description: Delete a program by ID.
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete program
      tags:
      - Programs
    put:
      consumes:
      - application/json
      description: Update an existing program.
      parameters:
      - description: Program ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated program
        in: body
        name: program
        required: true
        schema:
          $ref: '#/definitions/models.Program'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Program'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update program
      tags:
      - Programs
...
swagger: "2.0"
...